requirements.txt
Pillow~=9.4.0
numpy~=1.26.2
tensorflow~=2.15.0.post1
keras~=2.15.0
opencv-python~=4.8.1.78
python-dotenv~=1.0.0